ParamsBox moved to tss/params
This code is strictly tss-related so keeping it inside tss directory
makes a lot of sense.
